feat: created the usersDataSchema on an isolated file
