+++ frontend/src/zodSchemas/usersDataSchema.tsx
@@ -1,0 +1,11 @@
+import { z } from "zod"
+
+ const usersDataSchema = z.object({
+  firstName: z.string(),
+  lastName: z.string(),
+  email: z.string().email(),
+  password: z.string().min(8),
+  confirmPassword: z.string().min(8)
+})
+
+export default usersDataSchema
